updated the ui components
