+++ src/components/ui/badge/badge.module.css
@@ -1,0 +1,20 @@
+.root {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 9999px;
+  background-color: #edecec;
+  padding: 0.25rem 0.5rem;
+
+  @media (prefers-color-scheme: dark) {
+    background-color: #2c2c2c;
+  }
+}
+
+.title {
+  font-family: inherit;
+  color: var(--foreground, currentColor);
+  font-size: 0.75rem;
+  font-weight: 300;
+  margin: 0;
+}
