+++ src/components/ui/collapsible/collapsible.module.css
@@ -1,0 +1,66 @@
+.root {
+  background-color: var(--whitebg, #f5f5f5);
+  border-radius: 0.375rem;
+  box-shadow: 0 1px 2px 0 rgb(0 0 0 / 0.05);
+
+  @media (prefers-color-scheme: dark) {
+    background-color: #2c2c2c;
+  }
+}
+
+.trigger {
+  display: flex;
+  cursor: pointer;
+  align-items: center;
+  justify-content: space-between;
+  padding: 1rem;
+  font-weight: 300;
+  font-style: italic;
+  user-select: none;
+  font-size: 1rem;
+
+  @media (min-width: 640px) {
+    font-size: 1.125rem;
+  }
+
+  @media (min-width: 768px) {
+    font-size: 1.25rem;
+  }
+}
+
+.icon {
+  margin-left: 0.5rem;
+  transition-property: transform;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 300ms;
+
+  &.open {
+    transform: rotate(180deg);
+  }
+}
+
+.content {
+  overflow: hidden;
+  border-top: 1px solid #2c2c2c;
+  transition-property: all;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 500ms;
+
+  @media (prefers-color-scheme: dark) {
+    border-top-color: #edecec;
+  }
+
+  &.closed {
+    visibility: hidden;
+    max-height: 0;
+  }
+
+  &.open {
+    visibility: visible;
+    max-height: 24rem;
+  }
+}
+
+.content-inner {
+  padding: 0.5rem 1rem;
+}
