+++ src/components/ui/index.ts
@@ -1,0 +1,4 @@
+export { Button } from "./button/button";
+export { Badge } from "./badge/badge";
+export { Collapsible } from "./collapsible/collapsible";
+export { Toggle } from "./toggle/toggle";
