+++ src/components/ui/toggle/toggle.module.css
@@ -1,0 +1,28 @@
+.root {
+  display: flex;
+  gap: 0.5rem;
+
+  @media (min-width: 1024px) {
+    gap: 1rem;
+  }
+}
+
+.button {
+  color: var(--color-foreground);
+  display: flex;
+  cursor: pointer;
+  align-items: center;
+  justify-content: center;
+  transition-property: all;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 150ms;
+  background: none;
+  border: none;
+  padding: 0;
+  font-family: var(--font-body);
+  font-weight: 100;
+
+  &.active {
+    font-weight: 500;
+  }
+}
